Utility: don't have a billion Lorem Ipsum copies in every benchmark.
Reuse the file that's used for StringView benchmarks in Containers. It's
roughly the same size as the original embedded one was, so the
measurements don't change that much. On the other hand, we now copy it
10x and run 10x10 iterations on slices of it, instead of 100x1
iteration, which should hopefully lead to smaller measurement variance.
Plus it's easier to extend/shorten this way.