Utility: SSE2 variants of String::{lowercase,uppercase}().
Those are reverse-engineered from what GCC and Clang does with the
branchless variant in a baseline x86-64 Release build, i.e. with just
SSE2 enabled, and match it. It however doesn't do that for a
RelWithDebInfo build, there the scalar branchless code is roughly
equivalent to the naive implementation with a branch.