Containers: restrict StringView NEON code to AArch64.
The instructions it uses are only on AArch64, not on Armv7 etc.
Originally I wanted to add some differentiation between 32-bit and
64-bit to CORRADE_TARGET_NEON, CORRADE_ENABLE_NEON and Cpu::Neon, but
this seems like an obvious solution. especially given that majority of
ARM devices today is 64-bit anyway.

A similar thing was already done inside BitArrayView::count()
implementation, where a 64-bit popcnt instruction is used on x86 and on
32-bit builds it falls back to the scalar case, because 32-bit x86 CPUs
are simply too rare nowadays to actually care about.