Fix CORRADE_TARGET_SIND128 detection on Emscripten 3.1.23+.
The detection used __EMSCRIPTEN_major__ etc. to exclude Emscripten
older than 2.0.18 because it didn't have the actual intrinsics in Clang.
However, on Emscripten 3.1.23+ these are no longer implicitly defined,
which caused the detection to behave as if an old Emscripten was used,
never actually defining CORRADE_TARGET_SIMD128.

To fix that, <emscripten/version.h> has to be unconditionally included
by configure.h. which is a bit nasty. It'll get cleaned up again once
support for Emscripten pre-2.0.18 is dropped.