Utility: work around Clang and GCC picking wrong copy() for void views.
Today I built GCC 5, GCC 6, GCC 8, GCC 9 and GCC 10 from source just to
pinpoint the exact versions whare this was broken. Ugh. I still have no
idea what's wrong in these versions that makes it fail, and what all
versions of Clang and all versions of MSVC do differently than GCC 4.8
and GCC 11+ where this isn't a problem.

And then, THEN, when commiting and pushing everything, I discover this
is broken on GCC 11+ if ArrayView<void> and not ArrayView<const void> is
the first argument. GCC 4.8 still works as expected for it. The best
compiler out there. The industry peaked in 2014, it only went downhill
since.