Containers: preserve String AllocatedInit property in copies.
This makes more sense than relying just on the size alone -- in
particular, the main use case of AllocatedInit is to put strings into
growable arrays or vectors and be able to reference their data even if
the array is reallocated. That was indeed the case with Array (which
always only move-constructs during a reallocation), but may not be with
std::vector or 3rd party containers even though String moves are
noexcept etc, etc.

So now it's always allocated if the original was, even though SSO would
be possible and more efficient.