Containers: fix StaticArray NoInit with NoInit-constructible types.
The types all use an explicit constructor, but the test was written in a
stupid way and so it didn't test this at all. Now it does, and one has
to explicitly initialize the data as T[]{T{NoInit}, T{NoInit}...} in
order to make it work. Which is a copy initialization I think, so
theoretically it copies garbage over instead of not doing anything, but
the test seems to pass.