Containers: don't allow creating a StaticArray from a smaller array.
Sigh, this "helpful" C behavior makes me mad. Of course it caused an
extra time spent debugging because a function returned less values than
it should have.

In addition to the new array-taking constructors introduced in
eea549c2d7209d6e0a61bddd39ed920b63f0bac1, a static_assert() for this was
missing in one of the variadic constructors. That's a regression
introduced in 6b10c341269ff44a76b18a4e179c2af600c81ae1 due to the
constructor with static_assert() no longer being delegated to from the
other variant, sorry about that.

Unfortunately GCC 4.8 isn't able to figure out the size on its own, so
for it the original constructors are kept. Which means no size
assertions there, but nowadays projects should have GCC 4.8 only as one
of the targets (if at all), not the sole compiler being used, so the
misuses should get caught elsewhere anyway.