Containers: add a lightweight function wrapper.
Still got to look into actually optimizing like inlining where possible
or providing the C++17 way of passing function pointers via template
arguments, but wanted to have something working, documented and tested
first.

Surprisingly enough this works on all tested compilers including fancy
features like overload resolution. The only thing that doesn't work is
on MSVC 0215 overload resolution for function overloads taking different
Function types as an argument when a plain function reference is passed.
But it can be worked around with conversion to a function pointer first,
and nobody should be using this crap of a compiler for serious stuff
today anyway.