Containers: add a TODO for is_trivially_* workarounds on GCC 4.8.
Wrapping those in a std::integer_constant doesn't seem to be needed if
using them in template arguments instead of function arguments, so this
nasty helper may not be needed at all. Plus, template argument SFINAE
should be easier / faster to compile, some people say, so switching to
that is a good idea.