Containers: make use of all 64 bits for popcnt in StringView::count().
Compared to calling popcnt once for each 16-byte vector, this
accumulates masks of four vectors and then calls popcnt just once. Then
there's a less-than-four-vectors postamble that tries to put at least
two vectors together.

Even though this accumulation requires a bunch of extra shifts and
ORs, the speed improves from ~90 ns to ~70. The previous variant is
moved to the benchmark file for reference, similarly a bunch of variants
for the postamble handling are left in the benchmark file.