Containers: inline the SIMD StringView::find() helpers for readability.
Having them separated out was just because the original Rust code had it
that way, and after repeatedly re-reading this code, it doesn't actually
improve anything. The code is longer that way, the function call feels
like something *extremely* complicated is done there while it's
literally three lines, there's extra work for the optimizer, one has to
ensure this helper really gets inlined. And then there's three return
statements involved in every such invocation where it could be just one,
and it doesn't make it less bug-prone either because instead of checking
that `loadu` or `load` is correctly used, one now has to check that the
right function gets called and then that the actual function has the
right instruction inside. Just jumping all around the place, both for
the human and the compiler.