Utility: test benchmark-only String::lowercase() variants as well.
Similarly to what's just been done for StringView::count() and the
BitArrayView algos before, putting them into a header and loading
through a function pointer the same way as all other tested variants.

Makes it easy to ensure the implementation doesn't cut corners anywhere,
and can be ordered in the benchmark right next to the default variant it
compares to.