Utility: SSE4.1 variant of String::replaceAllInPlace().
First I tried a naive conditional replace, it was slower than the
unconditional version, and I was disappointed. But then I realized the
find() implementation was doing the conditional only every four vectors
and that actually *is* faster! So that's what goes here, and the
unconditional version is left only in tests for comparison.