Utility: fail Path::make() if given path exists but is not a directory.
Until now it silently returned true and didn't complain about anything,
now it prints an error message and returns false.

Unfortunately, as file type checking is unrealiable on certain
platforms, this extra check is done only on platforms where it *is*
reliable, so on the others it'll still silently return true. Which is
however better than the behavior suddenly breaking there, making the
function unable to create anything because nothing is ever reported as
being a directory.