Containers: drop the IsTrivially* helper structs for old GCC.
They were there because the compiler lacked the std::is_trivially_* type
traits, and the __has_trivial_*() builtins couldn't be used for SFINAE
in function signatures (but the wrapper struct could). However, when
making the Function wrapper in 86623c42a06a5d266a544cbb2ea0287aef37f99c
I discovered they can be used without the wrapper struct when used for
SFINAE in template arguments.

So that's the change. Plus, this way should also avoid some extra work
for the compiler in overload resolution. Or so people said, somewhere.