Convert remaining argument / return SFINAE into template argument SFINAE.
When these were a part of the function signature, they ended up being
encoded into the exported symbol. The worst offender was the Debug
operator<< for iterables, for which the mangled name was *half a
kilobyte*.

In total, for CorradeUtility-d.so, the output of `strings | grep
enable_if` went from 24 kB to 5 kB, the rest is std::swap overloads (for
which I need to de-STLify Arguments and Configuration), and
StridedArrayView slice(), which is a bit more complicated matter. The
(stripped) release build doesn't have any exported symbols containing
enable_if, as all those are inline.