PluginManager: avoid needless std::vector in AbstractManager.
This was I guess waiting for arrayRemove() to finally exist. The second
use is unfortunately not removable at the moment, as with dynamic
plugins on static builds growable arrays won't work in their current
shape.

The remaining std::vector and std::string uses are in PluginMetadata,
which will get reworked altogether.