PluginManager: port PluginMetadata away from STL.
The PluginMetadata class is now completely empty, a base of the internal
Plugin structure, which now contains all the data.

A backwards compatibility header is included for the name() getter, but
the remaining APIs returned std::vector<std::string>, and implicit
conversion of StringIterable to *that* is not available for obvious
footgun reasons. PluginMetadata are however a rather internal API that I
don't expect is used too heavily, so this change shouldn't be too
problematic I hope.