TestSuite: document how to deal with noise in benchmark measurements.
It was nowhere mentioned that increasing the repeat count might change
anything, now it's clear. What is obvious to some, might not be to
other. Similarly, the cpufreq command was only shown in the verbose
output, but never actually mentioned in the documentation, nor there was
any hint that passing `-v` would show it.

The /// is there because the shell command contains a */. Not great, but
I didn't feel like adding a shell file to doc/snippets just for this
alone.