Containers: allow array{Append,Insert}() to copy from the array itself.
But in the list-taking version only, as I don't want to pessimize the
single-value addition too much, and as this is (I think) more common for
ranges anyway (such as an allocator implemented on top of growable
arrays growing itself).

Single-item append / insert thus has a debug-only assert for this (thus
no penalty in release), and list-taking variants handle this case (where
the assumption is that usually there are many more items to copy to
amortize any extra overloead).

Funny thing is that std::vector just gives up and says this is undefined
behavior. The current understanding of C++ being unsafe is largely
caused by the STL not giving a fuck.