Utility: fix Algorithms.h with app-side CORRADE_NO_ASSERT overrides.
Before this change, if Corrade was compiled with the default setting
(i.e., assertions enabled) and the application was compiled with them
disabled, or vice versa, the build failed with a linker error because
the internal function signature was depending on this macro being
defined consistently between the library and the application.

Which isn't really great, as it should be possible to disable assert in just
some part of the chain, not all. And since this was the only such place
causing problems, all other signatures depending on the macro being
internal unexported symbols, it made sense to make it work.

A downside is that one extra unused pointer gets passed on no-assert
builds, but assuming the actual copy operation is substantial, the
overhead shouldn't be a problem.