doc: turn the CORRADE_CONSTEXPR* macros into plain constexpr in docs.
Ideally it'd show the standard in which it actually is constexpr, but
that needs a feature in m.css and has to wait.

I had to add defined(DOXYGEN_GENERATING_OUTPUT) to the definiton as
otherwise Doxygen would have those predefined already, which would skip
the branch, which would then make those disappear from the docs
completely.