package/ci: add a dedicated build for testing WASM SIMD128.
It wasn't tested because the other Emscripten version is too old and
Node.js is too ancient to even run that, and I'm skipping test builds
locally, which inevitably resulted in a regression.

Yeah and THEN, after trying to run on 2.0.25, it took a whole day of
suffering and trying to find a non-broken version. Such version doesn't
exist since then. 2.0.26 fixes a breakage in aligned allocations that
plagued 2.0.25, but it's broken with empty files. Everything since
2.0.26 is broken the same way, and then 3.1.3 fixes that by breaking
something else. Fuck me.