Containers: fix StringView::find(char) on WASM SIMD128.
Ultimately the refactor in 5b4e2a7dff9b087211d1f647ed112e7d1f0bc230 is to
blame for this bug as well, and unfortunately it went unnoticed even
though the tests were actually reproducing. The reason is that the
Emscripten tests take ages to build so I'm often just skipping them
locally, and the CI wasn't testing the WASM SIMD128 code because the
bundled Node.js in emsdk 2.0.25 was too old and I didn't bother.

In retrospect that wasn't a good idea, but after *a whole day* of extreme
suffering I managed to make emsdk working with Node.js 18 and even found
a version of emsdk that doesn't fail in brutal ways, only in one quite
localized and embarrassingly stupid way that can be XFAILed around.

Though this failure also manifested in quite spectacular ways, affecting
everything from the configuration parser to String benchmarks. Which
made me think it had to be something in Emscripten or Node.js that's
broken, leading to even more wasted time.