Containers: "fix" OOB access w/ negative strides on Emscripten 3.1.66+.
Still not sure whether this should be treated as a fix or as a
workaround. Tried to find all places where `data + i*stride` happens
with an unsigned `i` and cast it to a signed type, but as far as my
understanding goes, this should be completely fine, unless the compiler
does something crazy like calculating with a 64-bit temporary on a
32-bit system which then doesn't wrap around correctly.

*Especially* given that one of those places was a stride calculation
happening completely on a std::size_t, but still resultin in a correct
output afterwards.

Co-authored-by: Will Usher <will@willusher.io>