Utility: move std::pair Debug printer to DebugStl.h.
This avoids an #include <utility> in Debug.h. As basically all
std::pair, std::move() and std::forward() usage was eliminated, there's
not really any need to continue supporting this type by default, and
thus a breaking change like this should be fine.

With libstdc++, in C++11 it's mere 650 lines on top of <type_traits>
(that directly or indirectly get included by almost everything), it's
already 2.6k in C++23. That's already quite significant given it gets
pulled in almost everywhere

With libc++ (thus all Apple platforms, Android *and* Emscripten as well)
it's 4.5k in C++11 and 5.7k in C++23, which is a lot worse.