Utility: support redirecting Debug to Containers::String.
This FINALLY allows me to get rid of <sstream> and related STL compat
includes in almost all test code.

It's not great because it makes a heap allocation for the stream class
that contains the convoluted logic, but apparently that class is 376
bytes on libstdc++ (what the hell!!) so an allocation is actually
completely fine given the circumstances. Furthermore, reading through
stringstream internals it seems that the string gets internally
preallocated to 512 bytes (!!), so even with this massive 376 byte
allocation this is still less memory used overall if the output is under
100 chars, which often is.