Containers: move std::{pair,tuple} trivial tests to *StlTest.
Those rely on the type being implicitly available, which is at the
moment only because <sstream> is included. Once it won't be, the tests
would need to include <utility> / <tuple> which doesn't make sense
given there's another dedicated test for STL functionality.

Also turn those expected failures to actual XFAILs so it's clear that
STL is the worse thing.