Test: remove all use of <sstream> for Debug output testing.
The tests now all use redirection to a String, which means I can finally
drop also all STL compatibility headers. The only places where the
streams were kept is Debug tests themselves, as the stream output is
still the simpler way without too many moving parts, and various
TestSuite tests because there a std::ostream is used in an API to choose
between std::cout / cerr and a string for tests, i.e. it doesn't make
sense to add an extra overload *to the TestSuite library* just to make
its tests less heavier.

Alongside that I removed unneded usage of STL compatibility headers not
related to streams, and where they're still needed they now have a
comment why.

Compile-time-wise, comparing state of this commit with state before
cebfa6381e936f29b50a669dcbb49aaa73aaf55a (which removes <utility> usage
from Debug), on libstdc++ in C++11 mode the cumulative time across all
cored is reduced from ~3:53 to ~3:35 (so, in practice, about two seconds
on an 8-core machine). That's the most lightweight case, on the other end
of the extreme is libc++ in C++23 mode, where the time before is 7:08 and
time after is 5:35, so about 10 seconds saved on a 8-core machine. I
don't know about MSVC yet, where it's generally the worst, but the CIs
will tell me. All this is also the build with deprecated APIs enabled,
which pull in some STL compatibility headers, on a non-deprecated build
the difference may be larger.