modules: make corrade_add_test() link-depend on FILES on Emscripten.
On all other platforms (including Android, where those files get copied
to the device when running given test) these are just a runtime
dependency and is enough for CTest to just yell if they don't exist. But
on Emscripten they are *embedded* into the executable, and thus when
they change it has to be relinked.