Containers: fix StridedArrayView::slice() for non-overloaded member fns.
Finally. It took me three years to realize a solution that is actually
possible to implement -- basically keeping the original overloads, which
restrict themselves to const functions only on const types, and
non-const functions only on non-const types, and only adding another
overload that takes an arbitrary type that's subsequently converted to
a member function pointer. Which will work only for non-overloaded
functions, thus making the variants complete.