Containers: work around an ICE on MSVC when slicing to an array member.
Slicing to a member function returning an array reference, in
particular. Which is what I wanted to make use of in Magnum. in
particular code like this:

    Containers::StridedArrayView1D<Vector3s> a;
    Containers::StridedArrayView2D<Short> b = a.slice(&Vector3s::data);

Fortunately the workaround is quite clean and compact, although both GCC
and Clang rightfully complain that I'm doing something fishy if I enable
it on those.