Containers: use std::is_same instead of std::is_convertible for clarity.
This was apparently copypied from StridedArrayView, where it makes
sense, as e.g. StridedArrayView<Vector4> can be constructed from
ArrayView<Color4> assuming Color4 is a subclass of Vector4 and has the
same size. The same check works for constructing a const view from a
mutable one.

But in case of StridedBitArrayView the type is always either just char
or const char, so the check is unnecessarily broad, and should be really
checking for just const correctness but the type being the same.