Containers: don't actually need to use the argument SFINAE after all.
The type is known, even though the syntax is a bit strange, so turn that
into a template argument SFINAE like elsewwhere. This was the final
instance of enable_if as part of a signature, down from ~5 kB. The rest
is due to remaining STL use such as std::string or std::tuple.

    $ strings libCorradeUtility-d.so | grep enable_if | wc -c
    3582

Across all Magnum libraries this causes a reduction from ~18 kB to just
this.

    $ strings libMagnum*-d.so | grep enable_if | wc -c
    4103