Containers: allow creating views on STL containers of derived type.
The use case was with Magnum Ui library LineLayer, where a function
accepts a StridedArrayView1D<Vector4> but application code uses
std::vector<Color4>, where Color4 is a specialized subclass of Vector4.
Such conversion already worked among Corrade view types themselves, but
not when an external type was involved. Now it works for those too.

Also expanding the existing tests to verify that it's not accidentally
possible to create mutable views from const types now, due to the
additional complexity in the code.