Containers: prevent mangling incompatibility between GCC and Clang.
Hah, I didn't get ABI breakages between the two in quite a while. This
is caused by the changes in 78777968b13a8be69d317cd6d0a09650e79bf11b and
these three occurences are affected because they're a SFINAE-disabled
member that's only enabled on a mutable view, and the implementation is
deinlined in the cpp file. In all other cases the implementation is in
the header so even if the mangled name is different, it doesn't really
matter as the compiler makes an inline copy anyway.

So I'm reverting those to what was there before, i.e. a template type
parameter rather than template value parameter, because it's vital to
have it possible to build with one compiler (such as a distro package)
but then allow people to use the other compiler if they want.