Utility: redo deprecated String::view{Begins,Ends}With() with StringView.
There should be no usage difference compared to the previous
implementation as an ArrayView is implicitly convertible to a
StringView, but now we don't have the same code in several places
anymore and don't need to rely on ArrayView being included.

Also fix the deprecation message to mention the correct name.