Utility: make Json{Object,Array}Iterator and View without templates.
It behaves slightly differently for objects and arrays, which makes it a
bit confusing if it's all merged into the same type. Plus I want to be
able to use the JsonView and JsonIterator names for iteration directly
on tokens. Having those as templates prevents that.

This is a minor breaking change, as the JsonView and JsonIterator types
no longer exist. But I assume all user code used the JsonObjectView /
JsonArrayView typedefs, and the JsonIterator type was hidden and used
only in range-for loops.