Utility: turn JsonToken into an opaque reference type.
With this it means user code no longer needs to pay extreme attention to
not accidentally copy the JsonToken references returned by the Json
APIs. It now stores a pointer to the token array and an index into that
array and can be treated as a regular value type with no surprising edge
cases. Plus it's now also just the size of two pointers instead of three
or more and thus can be passed by value in all arguments as well.

Besides that it now also allows for further decoupling of the public
interface and internal data layout, along with additional safety checks
that were not possible before. Those will appear in the next commits.

This should be for the most part a fully source-compatible change, as
all existing code that saves returned JsonToken values into const& will
now make use of C++'s reference lifetime extension feature. Which is why
both the test code and the snippets still use const&, to verify it all
compiles and works like before.