Utility: unify internal JSON token representation between 32 and 64 bits.
Even after changing parsed string references to indices instead of
pointers in a previous commit I couldn't just reuse the 32-bit
representation to 64-bit because the 32-bit one was not using enough
bits to represent child count and such. The new representation does,
using 48 bits from the mantissa, which should be plenty.

A downside is that I had to make the type bit packing more optimal,
which means queries like JsonToken::type(), parsedType() and isParsed()
are now more complicated than they used to be, and thus likely becoming
a bottleneck. On the other hand, the bit encoding for types is now more
robust, meaning that instead of all the branching in type(),
parsedType() etc., the assertions can be simplified to just a single
bitmask expression. Which makes them even simpler than before, and the
type(), parsedType() etc. won't be used in hot paths at all anymore.
That change is not part of this commit and will be done separately.

Ultimately, while things get more complicated in certain parts, it's
overall less code because there isn't any preprocessor branching
anymore, and updating any code no longer means inventing the same thing
twice. Yay.