Utility: macros for overriding behavior for all asserts at once.
The assertion macros grew organically over time and now if even a
slightly different behavior is intended such as exiting instead of
aborting it's just too annoying to override them one by one. For that
there's now CORRADE_ASSERT_ABORT(), which changes just the aborting
behavior and CORRADE_ASSERT_MESSAGE_ABORT() which can modify also the
message printing.

Finally, with CORRADE_ASSERT_INCLUDE it's possible to pass a header file
that gets included every time Assert.h is included, to allow for easy
project-wide customization without having to include such a header
unconditionally to each and every compilation unit (which can be
prohibitively expensive) or deal with compiler-specific ways of
specifying an include file. For example on GCC it's -include I thin, but
I have no idea if it needs to be added first or last to take the -I
paths into account, etc. etc. This macro circumvents that and reduces it
to "just" the pain of specifying a double-quoted path somehow.