Utility: have to include CORRADE_ASSERT_INCLUDE outside of the guard.
Otherwise, if the user-provided header includes Assert.h, either
directly or transitively, it'll be skipped as a whole because the header
guard is already defined, leading to none of the assertion macros being
defined.

This is, I suppose, also one of the few cases where #pragma once
wouldn't work -- it would simply not include the Assert.h from within
CORRADE_ASSERT_INCLUDE, leading to the same problem.