Deprecate all DefaultInit initialization.
The DefaultInit tag, which maps to C++'s default initialization, where
trivial types are left uninitialized and default constructor is called
otherwise, was discouraged for quite a while already, suggesting to use
ValueInit or NoInit instead to make the choice explicit and avoid
accidents.

Furthermore, because of the default initialization being
prone to errors, new APIs such as String or BitArray didn't have
DefaultInit variants at all. Now I got to a point where I want to add
ValueInit variants of some NoInit APIs but don't want to deal with the
discouraged DefaultInit, so this finishes that process and marks all
existing uses of DefaultInit as deprecated.

The DefaultInit tag itself *isn't* deprecated, as it's used to mean "use
reasonable defaults for this value" in other contexts, in particular
Magnum's uniform structs. There it makes sense, and the sooner we forget
about "default initialization" meaning the weird C++ thing, the better.