doc: clarify behavior with Global string literals.
A lot of the PluginManager code was apparently written before I realized
that Global string literals stored in a Containers::Stirng have to be
NullTerminated as well, as otherwise the internal invariants would break
there. So update the docs to match what's actually done by
String::nullTerminatedGlobalView(), or remove the guarantees altogether
if they no longer hold.

And update the doc formatting to not show the redundant enum prefix.