Utility: conditionally enable Windows-specific Path behavior on Emscripten.
If a command line executable such as corrade-rc (or, say,
magnum-imageconverter from Magnum) is used on Windows, it should
correctly deal with Windows specifics for paths, namely drive letters
and backslashes. So far Emscripten was treated as Unix and no such
special-case was done, making it impossible to use Emscripten-built
corrade-rc when compiling on Windows.

Because an Emscripten binary can be built on Linux and then ran on
Windows (which is the case especially with the Magnum utils, less so
with corrade-rc), we have to do the detection at runtime. In turn it
means that a call to Path::fromNativeSeparators() or toNativeSeparators()
will produce some JS garbage as a side effect. For those two APIs it
shouldn't matter much as they're likely only being called when
processing command-line arguments or calling 3rd party APIs that require
backlashes on Windows, not being regularly used with thousands of
strings.

In case of Path::join() the logic is that Windows is queried through
JS only if the path looks like having a Windows drive letter. That
should also reduce the overhead for >90% of use cases. I added a
benchmark and with this shorthand evaluation, the call to JS makes
join() roughly twice as slow. Which isn't too bad, and given
that filesystem operations are quite rare in JS -- with majority use
cases being browsers that don't access the filesystem in any way -- it
should be fine.

Another option here could be to cache the result in a static
function-local variable (so, three static variables, one in each API).
But that may introduce thread-local variables and/or mutexes to ensure a
single initialization, likely inflating code size, and there's a slight
chance that such extra bloat won't get DCEd if these functions aren't
used. I think code size is a bigger priority than minor speed increase
for a "1% use case" so I'm leaving it like this.