Containers: disallow use of set(bool) and reset(bool) on bit arrays.
After not working with bit arrays for a while, I wrote the following in
an attempt to set all bits to false:

    array.set(false);

and then wasted quite some time figuring out why some bits were still
set, because what this did was setting bit 0 to true, not all bits to
false. While I could add a bool overload that treats this the same way
as setAll() / resetAll(), it's less error-prone to just delete the bool
overload to force a compile error if a bool is used instead.

(And of course this ultimately wasn't as simple as adding a
`void set(bool) = delete;` overload, because C++ and the implicit type
conversion madness. Ugh.)