Main: actually use wWinMain() in MinGW GUI apps, not wmain().
Ugh, this was so awful. Two days of investigating totally cursed side
effect only to realize that *MinGW itself* defines wmain() that then
delegates to wWinMain(), and when CorradeMain (which contains both) is
linked, the wmain() from CorradeMain gets used instead of the one from
the MinGW runtime, leading to wWinMain() never called.

This is a potentially breaking change for all MinGW users that build GUI
apps, sorry, could not be done any differently. If you have a copy of
FindCorrade in your project, update it to the latest version, and
everything should continue to work as before. If you use CMake and don't
have a copy of FindCorrade, you don't need to do anything. If you don't
use CMake, you now have two different libraries to link to, along with
extra flags.