Utility: prefer standard attribute for CORRADE_DEPRECATED*() if possible.
In other words, when compiling as C++14, C++17 or later, it should use
the standard [[deprecated]] attribute and not a compiler-specific
extension. This makes the code more future-proof and also enables the
macros to potentially work even on unknown compilers (i.e., a compiler
that is neither GCC, Clang or MSVC, if that's even a thing today).

Unfortunately, for CORRADE_DEPRECATED() the standard attribute placment
differs from the placement where the compiler-specific attributes were
used, so to avoid breaking existing code there I'm staying with the
original version for now. Once I manage to make a tagged release and
most deprecated typedefs are gone, I'll look at this again.

Also, because the macros now behave differently based on the C++
standard used, the DeprecationMacrosTest is now compiled for C++14 and
C++17 as well, verifying that the exact same code compiles on all three
standards.